Bot Control API/requirements.txt
fastapi==0.95.2
python-jose==3.3.0
python-multipart==0.0.6
uvicorn==0.22.0
passlib 
bcrypt==4.0.1
sqlalchemy==2.0.18
python-dotenv>=1.0.0

docker
docker-compose